fix(boundarypackage): do not print rates to list file table if ibound <= 0 (#783)
* The number of rows in the bound package table printed to the list file was not consistent with the number of rows written to the table.  The table writer was modified so that an entry is not written if ibound is less than or equal to zero.  This is consistent with mf2005.
* Added new recharge test that mixes irch and idomain
* Minor tweaks to mf6bmi.f90 and test_gwf_npf01_75x75.py

+++ srcbmi/mf6bmi.f90
@@ -439,7 +439,7 @@ contains
     use MemorySetHandlerModule, only: on_memory_set
     ! -- dummy variables
     character(kind=c_char), intent(in) :: c_var_address(*) !< memory address string of the variable
-    type(c_ptr), intent(in) :: c_arr_ptr                    !< pointer to the double precision array
+    type(c_ptr), intent(in) :: c_arr_ptr                    !< pointer to the integer array
     integer :: bmi_status                                   !< BMI status code
     ! -- local variables
     character(len=LENMEMPATH) :: mem_path
